requirements.txt
groq==0.31.0
llama-index-core==0.13.2
llama-index-readers-file==0.5.1
llama-index-embeddings-huggingface==0.6.0
llama-index-llms-groq==0.4.0
llama-index-llms-vertex==0.6.0
llama-index-vector-stores-chroma==0.5.0
tiktoken==0.11.0
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
streamlit==1.37.1
requests==2.32.3
numpy<2
torch==2.2.2
transformers==4.44.2
sentence-transformers==2.7.0
python-pptx==0.6.23
evidently==0.5.0
pinecone-client==6.0.0
redis==6.4.0
llama-index-vector-stores-pinecone==0.7.0
retry==0.9.2